started 14.1

+++ AP Lesson 14/Location.java
@@ -1,5 +1,5 @@
 public interface Location
 {
    double[] getLoc();
-   int getID();
+   double getID();
 }
